[~] used TvLazyHorizontalGrid and code optimizations
