+++ tvApp/src/main/java/band/effective/office/tv/screen/eventStory/EventStoryViewModel.kt
@@ -172,7 +172,6 @@ class EventStoryViewModel @Inject constructor(
                     .toUi(employees)
                     .take(countUsersToShow)
                     .sortedByDescending { it.totalSeconds }
-                    .toTwoColumns(countUsersToShow)
             ) as StoryModel
         )
 
@@ -186,7 +185,6 @@ class EventStoryViewModel @Inject constructor(
                     .toUi(employees)
                     .take(countUsersToShow)
                     .sortedByDescending { it.score }
-                    .toTwoColumns(countUsersToShow)
             ) as StoryModel
         )
